Move the connect.Request to daemon.Request.
