+++ pkg/client/cli/util/init_command.go
@@ -6,7 +6,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/ann"
-	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/connect"
+	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/daemon"
 )
 
 type userDaemonKey struct{}
@@ -51,8 +51,8 @@ func CommandInitializer(cmd *cobra.Command) (err error) {
 		as[ann.VersionCheck] = ann.Required
 	}
 	if v := as[ann.UserDaemon]; v == ann.Optional || v == ann.Required {
-		if cr := connect.GetRequest(ctx); cr == nil {
-			ctx = connect.WithDefaultRequest(ctx, cmd)
+		if cr := daemon.GetRequest(ctx); cr == nil {
+			ctx = daemon.WithDefaultRequest(ctx, cmd)
 		}
 		if ctx, err = ensureUserDaemon(ctx, v == ann.Required); err != nil {
 			if v == ann.Optional && err == ErrNoUserDaemon {
